Fixes CSS Location for Composer
When using Composer to install the module, the module directory is called silverstripe-invisible-spam-protection and not InvisibleSpamProtection. This was causing the CSS to fail to be included when installed via Composer. 

This update first checks if silverstripe-invisible-spam-protection exists and uses the CSS there if so, if not, it reverts to the CSS inInvisibleSpamProtection.